Ana/requirements.txt
requests==2.9.1
django==1.9.1
djangorestframework==3.3.2
django-rest-swagger==0.3.5